anstract strings into separate file
