+++ constants/index.js
@@ -1,0 +1,12 @@
+// if you want to use all the same strings throughout the app,
+// then I suggest these stay in alphabetical order so that you can
+// more easily see everything when you go to make a new string
+const APP_STRINGS = {
+  briefSummary: "Please Give Us a Brief Summary of Your Work Experience",
+  choose: "Choose",
+  openModal: "Open modal",
+  setWorkExperience: "Please Set Work Experience",
+  workExperienceSet: "Work Experience Set",
+};
+
+export const { APP_STRINGS };
